+++ Elder.ly/MasterViewController.swift
@@ -102,6 +102,17 @@ class MasterViewController: UIViewController, UITableViewDelegate, UITableViewDa
         let controller = AddEditViewController(nibName: nil, bundle: nil)
         self.navigationController?.pushViewController(controller, animated: false)
     }
+    
+    // UNWIND
+    @IBAction func prepareForUnwind(segue: UIStoryboardSegue) {
+        print("prepareForUnwind")
+    }
+    
+    override func unwind(for unwindSegue: UIStoryboardSegue, towardsViewController subsequentVC: UIViewController) {
+        print("unwind")
+        let segue = TransitionToLeftSegue(identifier: unwindSegue.identifier, source: unwindSegue.source, destination: unwindSegue.destination)
+        segue.perform()
+    }
 
     // MARK: - Segues
