+++ Elder.ly/MenuViewController.swift
@@ -24,29 +24,4 @@ class MenuViewController: UIViewController {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
     }
-    @IBAction func closePressed(_ sender: Any) {
-        self.dismiss(animated: true)
-    }
-}
-
-extension UIViewController {
-    
-    func presentMenu(_ viewControllerToPresent: UIViewController) {
-        let transition = CATransition()
-        transition.duration = 0.25
-        transition.type = kCATransitionPush
-        transition.subtype = kCATransitionFromLeft
-        self.view.window!.layer.add(transition, forKey: kCATransition)
-        present(viewControllerToPresent, animated: false)
-    }
-    
-    func dismissMenu() {
-        let transition = CATransition()
-        transition.duration = 0.25
-        transition.type = kCATransitionPush
-        transition.subtype = kCATransitionFromRight
-        self.view.window!.layer.add(transition, forKey: kCATransition)
-        
-        dismiss(animated: false)
-    }
 }
