--- Elder.ly/TransitionFromLeftSegue.swift
@@ -1,27 +1,0 @@
-//
-//  TransitionFromLeftSegue.swift
-//  Elder.ly
-//
-//  Created by Nicolas Vergoz on 01/12/2017.
-//  Copyright © 2017 Old Mojito. All rights reserved.
-//
-
-import UIKit
-
-class TransitionFromLeftSegue: UIStoryboardSegue {
-    
-    override func perform() {
-        
-        let src = self.source
-        let dst = self.destination
-        let transition: CATransition = CATransition()
-        let timeFunc : CAMediaTimingFunction = CAMediaTimingFunction(name: kCAMediaTimingFunctionEaseInEaseOut)
-        transition.duration = 0.3
-        transition.timingFunction = timeFunc
-        transition.type = kCATransitionPush
-        transition.subtype = kCATransitionFromRight
-        
-        src.view.window?.layer.add(transition, forKey: nil)
-        src.present(dst, animated: false, completion: nil)
-    }
-}
